Decided to remove the CS presentation title and body images. ------------------------------------------------------------
They are too heavy and I have a nice css instead.

All the presentations that use the images as background are dated, and should be revised anyway.
The git presentation in this commit is an example of what should be changed in order to use the css instead of the background images.

Also - updated reveal to latest.

Signed-off-by: dave <ddovrat@cs.technion.ac.il>